Added wait time and IsCrucial field
Added wait time for actions to allow time for action completion and I added an IsCrucial for the assertion.  If IsCrucial is true and the assertion fails, the test stops.  If IsCrucial is false and the test fails, the test keeps running.  IsCrucial should be set to true for all Navigation actions direct or indirect to ensure that you are on the correct page before attempting to check page elements.