Added screen shot limit and turned it into an App
Added screen shot limit and turned this into an Application that can be run from the command line.   You have to include all jars in the command path.  Now, it checks if the config file exists in the default location and if not, requests the path of the configuration file.  If a valid path is provided, the program executes and outputs to the terminal or command window.