Help file overwrite each time
Help file will be created each time the app is run as the first step.   Added the configuration file format to the help file and expanded upon the test file help.  Added Maximum number of screenshots configuration variable to the configuration file to limit the number of screen shots that will be taken in total but this does not apply to screenshots taken due to errors.