Added not equal, persisted string and ordered files
Added not equal comparison to check that strings are not equal.  Added ability to persist a string for a test file.  Added the ability to order test files using the number at the end of the file along with a new configuration field that determines if that sort will be used.