ConfigurableTester version 2.0 Initial Upload
With this new version, the test file is now xml based.  This is still under development.   The xml test format can be found in the Tests folder.   All Test files that end with -Test.xml have been tested and each test step works as it should.  The files that start with Gary are a collection of test steps that will be used to create the future help file that outlines how each test step should be created in the xml test file.