Updated help file and...
Updated help file and wrote a new alphanumeric  sort routine as the test files must be entered in numeric order so the numeric portion of the sort was removed and alphabetically sorting is all that is left.    I also updated the Project_Setup.txt file outlining what now needs to be done for setup, but it requires further editing to be complete.