Added Tear Down method
Added Tear Down method to ensure it is always run when executing withing the environment so that the browser always shuts down.  Cleaned up the Standalone Application output so that when the app is run outside of the API the ansi codes do not write to the screen.  Fixed some help items removing legacy commands in the updated help file.