ScreenShot Update for consistent sizing
Added functionality to allow for ScreenShots to have configurable sizing, and added content area sizing to the output to allow the tester to run a test, see the different sizes of screenshots, take the Max Width and Max Height settings from all screenshots and apply them as arguments in the ScreenShot test steps to create consistently sized images that can be compared against one another and have a pixel percentage comparison done.   Also changed the iFrame argument reordering functionality to allow for sorting any number of arguments.